Add clean fail as todo to Perldl2/README
Devel::REPL is very powerful, but the stack trace one gets
if something is missing is *horrible*.  This was a note to
add better error handling with graceful degradation rather
than just dying.  e.g., if something is missing, don't load
it and continue if possible...with a warning message.