Fix pdl2 fallback to perldl on win32 systems
This is sf.net #3056142.  The problem appears
to be with running exec on win32 and the weak
cmd shell environment.  The file handles of
the original pdl2 script did not seem to get
passed to the newly exec-d perldl but opened
again instead.  That resulted in interleaved
IO between the two instances.

The fix was to replace the exec call by a
do filename call instead so perl handles the
new program rather than the shell getting
involved.