Add $PDL::PREFIX_RE to perldl and pdl2
This automatically strips prefixes matching the PDL shell
prompts so that cutting and pasting from the example output
in the PDL Book will just work.  In addition to adding that
to perldl (already was in pdl2), we now strip out the
leading prefix if found from the entries in the history
list.